EchoPrismAgent/requirements.txt
# EchoPrism Agent Service — merge of backend + agent deps (minus unused)
fastapi>=0.124.1,<1.0.0
starlette>=0.49.1,<1.0.0
uvicorn[standard]==0.34.0
firebase-admin==6.6.0
google-cloud-firestore>=2.19.0
google-cloud-storage==2.18.2
python-multipart==0.0.18
python-dotenv==1.0.1
google-genai>=1.0.0
Pillow>=10.0.0
httpx>=0.27.0
websockets>=14.0
